move tests shared functions to /tests/utils
