+++ test/checkChallenge.test.js
@@ -1,16 +1,16 @@
 // vim: tw=120 softtabstop=4 shiftwidth=4
 
 const assert = require('assert');
+const https = require('https');
+const sinon = require('sinon');
 
 let config;
 let connection;
+
 const console = require('../console').console;
 
-const sinon = require('sinon');
-
-const https = require('https');
-
-const { FakeSocket, FakeAPI, base_challenge } = require('./test')
+const { base_challenge } = require('./utils/base_challenge');
+const { FakeSocket, FakeAPI } = require('./test')
 
 function stub_console() {
     sinon.stub(console, 'log');
@@ -22,19 +22,6 @@ function requireUncached(module) {
   return require(module);
 }
 
-function getNewConfig() {
-  const config = requireUncached('../config');
-
-  config.DEBUG = true;
-  config.apikey = 'deadbeef';
-  config.host = 'test';
-  config.port = 80;
-  config.username = 'testbot';
-
-  config.bot_command = ['gtp-program', '--argument'];
-  return config; 
-}
-
 afterEach(function () {
     sinon.restore();
 });
@@ -44,7 +31,10 @@ describe('Challenges', () => {
   let conn;
  
   beforeEach(function() {
-    config = getNewConfig();
+    config = requireUncached('../config');
+    const { assignNewConfig } = require('./utils/assignNewConfig.js');
+    assignNewConfig(config);
+    
     connection = requireUncached('../connection');
 
     stub_console();
